Refactor store operations methods into the storeops module
