+++ src/index.js
@@ -8,6 +8,7 @@ import Routes from './routes';
 import rootReducer from './reducers/root';
 import * as ApiComms from './modules/apicomms';
 import * as Actions from './actions/actions';
+import getPokemonsInStore from './modules/storeops';
 import AppStyles from './components/styles/app.module.css';
 
 const store = createStore(
@@ -15,34 +16,10 @@ const store = createStore(
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
 );
 
-const getPokemonsInStore = async pokemon => {
-  const pokemonInfo = await ApiComms.getOnePokemonInfo(pokemon.name);
-  const {
-    id,
-    name,
-    sprites: { front_default: imagesm },
-    sprites: { other: { dream_world: { front_default: imagelg } } },
-    types,
-    weight,
-    height,
-  } = pokemonInfo;
-
-  const onePokemon = {
-    id,
-    name,
-    imagesm,
-    imagelg,
-    types,
-    weight,
-    height,
-  };
-  store.dispatch(Actions.addPokemon(onePokemon));
-};
-
 const dataInit = async () => {
   const { results: pokemons } = await ApiComms.getPokemons(1);
   pokemons.forEach(pokemon => {
-    getPokemonsInStore(pokemon);
+    getPokemonsInStore(pokemon, store.dispatch);
   });
 };
